Update required PyTorch version and torch.var arguments
The restriction to PyTorch 2.8.0 applied only to an early iteration of the code and is no longer relevant. 

According to the documentation, the "unbiased" argument in torch.var was renamed to "correction" beginning with PyTorch 2.0; therefore, impose a version minimum requirement of 2.0.0. 

Tested the module with the latest version 2.9.1, and other versions >= 2.0.0 worked fine.